feat: redesign ApyChart and add ProtocolHeader for improved APY display
